requirements.txt
google-adk>=1.0.0
fastmcp>=0.1.0
uvicorn>=0.23.2
fastapi>=0.100.0
pandas>=2.0.0
google-cloud-bigquery>=3.11.0
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf>=6.1.3
db-dtypes>=1.0.0
tabulate
langchain-google-vertexai